+++ nyc.config.js
@@ -1,7 +1,7 @@
 /*
  * @Author: Tan Xuan
  * @Date: 2020-10-29 10:51:00
- * @LastEditTime: 2020-10-30 17:18:13
+ * @LastEditTime: 2020-11-18 12:23:24
  * @LastEditors: Tan Xuan
  * @Description:
  */
@@ -13,7 +13,7 @@ module.exports = {
   include: ["src/**/*.js"],
   exclude: [
     // 'src/modules/scroll.js',
-    // 'src/modules/wechat.js',
+    'src/modules/wechat.js',
     // 'src/modules/cookie.js',
     "src/config/*.js",
     "src/index.js",