+++ rollup.config.js
@@ -3,6 +3,8 @@ import commonjs from '@rollup/plugin-commonjs';
 import json from '@rollup/plugin-json';
 import { eslint } from 'rollup-plugin-eslint';
 import friendlyFormatter from 'eslint-friendly-formatter';
+import { uglify } from 'rollup-plugin-uglify';
+import babel from 'rollup-plugin-babel';
 
 export default {
   input: ['src/index.js'],
@@ -17,15 +19,20 @@ export default {
   plugins: [
     eslint({
       include: ['src/**'],
+      exclude: ['node_modules/**'],
       formatter: friendlyFormatter,
       throwOnError: true,
       fix: true,
+    }),
+    babel({
+      exclude: ['node_modules/**'],
+      runtimeHelpers: true
     }),
     resolve({
       mainFields: ['jsnext', 'module', 'main'],
     }),
     json(),
-    // babel({ runtimeHelpers: true }),
     commonjs(),
+    uglify()
   ],
 };