+++ src/index.js
@@ -1,7 +1,7 @@
 /*
  * @Author: Tan Xuan
  * @Date: 2020-09-24 11:21:11
- * @LastEditTime: 2020-11-17 12:52:28
+ * @LastEditTime: 2020-11-18 14:14:31
  * @LastEditors: Tan Xuan
  * @Description: 出口文件
  */
@@ -24,7 +24,22 @@ import {
 } from './modules/scroll';
 import { isWeChat, isAndroid, isIOS, isMobile } from './modules/userAgent';
 import { checkBank, checkEmail, checkId, checkPhone } from './modules/verify';
-import { weChatSDK } from './modules/wechat';
+import { weChatShare } from './modules/wechat';
+import { now, delay, debounce, throttle } from './modules/tool';
+import {
+  isObject,
+  isNull,
+  isUndefined,
+  isBoolean,
+  has,
+  isString,
+  isNumber,
+  isDate,
+  isRegExp,
+  isArray,
+  isFunction,
+  isSymbol
+} from './modules/tag';
 
 module.exports = {
   setCookie,
@@ -52,5 +67,66 @@ module.exports = {
   checkEmail,
   checkId,
   checkPhone,
-  weChatSDK,
+  weChatShare,
+  now,
+  delay,
+  debounce,
+  throttle,
+  isObject,
+  isNull,
+  isUndefined,
+  isBoolean,
+  has,
+  isString,
+  isNumber,
+  isDate,
+  isRegExp,
+  isArray,
+  isFunction,
+  isSymbol
+};
+
+export default {
+  setCookie,
+  getCookie,
+  deleteCookie,
+  formatSecondToTime,
+  formatDate,
+  calcAge,
+  formatPrice,
+  formatNumberToChina,
+  formatBankNumber,
+  randomNumber,
+  getScrollTop,
+  setScrollTop,
+  getRootScrollTop,
+  setRootScrollTop,
+  getElementTop,
+  getVisibleHeight,
+  getVisibleTop,
+  isWeChat,
+  isAndroid,
+  isIOS,
+  isMobile,
+  checkBank,
+  checkEmail,
+  checkId,
+  checkPhone,
+  weChatShare,
+  now,
+  delay,
+  debounce,
+  throttle,
+  isObject,
+  isNull,
+  isUndefined,
+  isBoolean,
+  has,
+  isString,
+  isNumber,
+  isDate,
+  isRegExp,
+  isArray,
+  isFunction,
+  isSymbol
 };