+++ changelog.md
@@ -1,11 +1,21 @@
+## 1.0.7
+
+- 由于 weChatSDK 只做微信分享，更名为 weChatShare
+- 新增 `now, delay, debounce, throttle` 方法
+- 新增 `isObject, isNull, isUndefined, isBoolean, has, isString, isNumber, isDate, isRegExp, isSymbol, isArray, isFunction` 等方法
+
 ## 1.0.6
- - 修改导出的wechatSDK变更为 weChatSDK, 并修改了参数
+
+- 修改导出的 wechatSDK 变更为 weChatSDK, 并修改了参数
 
 ## 1.0.5
- - 调整导出方式
+
+- 调整导出方式
 
 ## 1.0.2
- - add API document
+
+- add API document
 
 ## 1.0.0
- - publish
+
+- publish